% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{.AdjMat}
\alias{.AdjMat}
\title{Internal function to build the adjacency matrix.}
\usage{
.AdjMat(Network, decay = 0, Today = Sys.Date())
}
\arguments{
\item{Network}{Network dataset.}

\item{decay}{Decay factor.}

\item{Today}{Date.}
}
\value{
An adjacency/weight matrix
}
\description{
Internal function to build the adjacency matrix.
}
