% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{.QueryVector}
\alias{.QueryVector}
\title{Internal function to set up query vector}
\usage{
.QueryVector(fraudMat, decay, Today)
}
\arguments{
\item{fraudMat}{Data frame with binary indicator for fraudulent claims}

\item{decay}{Decay parameter.}

\item{Today}{Date}
}
\value{
Query vector
}
\description{
Internal function to set up query vector
}
