% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{.SNMM}
\alias{.SNMM}
\title{Internal function to set up symmetrically normalized matrix}
\usage{
.SNMM(aMat)
}
\arguments{
\item{aMat}{Adjacency matrix.}
}
\value{
symmetrically normalized matrix
}
\description{
Internal function to set up symmetrically normalized matrix
}
