% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{Homophily}
\alias{Homophily}
\title{Function to examine whether the network exhibits signs of homophily}
\usage{
Homophily(SimObj)
}
\arguments{
\item{SimObj}{an object of class \code{sfnData}, resulting from \code{\link{sfnGenerator}}}
}
\value{
A vector with the dyadicity and heterophilicity.
}
\description{
This function take an object of class sfnData as input and computes the dyadicity and heterophilicity.
}
\examples{
\donttest{
data("SimObj")
Homophily(SimObj)
}
}
\seealso{
[sfnGenerator()]
}
