\name{iFraudSimulator}
\alias{iFraudSimulator-package}
\docType{package}
\title{General information on the \code{iFraudSimulator package}
}
\description{
  This package contains the simulation machine engineered to create synthetic data with a network structure and available covariates similar to the real life insurance
  fraud data set analyzed in Óskarsdóttir et al. (2022). The user has control over several data-generating mechanisms. We can specify the total number of
  policyholders and parties, the desired level of imbalance and the (effect size of the) features in the fraud generating model. As such, the simulation engine enables
  researchers and practitioners to examine several methodological challenges as well as to test their (development strategy of) insurance fraud detection models in a
  range of different settings. Moreover, large synthetic data sets can be generated to evaluate the predictive performance of (advanced) machine learning techniques.
}
  \references{
    Campo, Bavo D.C., and Antonio, Katrien (2023). An engine to simulate insurance fraud network data. \emph{arXiv:}, available at

    Óskarsdóttir, M., Ahmed, W., Antonio, K., Baesens, B., Dendievel, R., Donas, T. and Reynkens, T. (2022). Social network analytics for supervised fraud detection in insurance. \emph{Risk Analysis} \bold{42}(8), 1872–1890.
}
