% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{normalize}
\alias{normalize}
\title{Min-max scaling to normalize a vector}
\usage{
normalize(x, Type = c("1", "2"))
}
\arguments{
\item{x}{an object of type vector.}

\item{Type}{a character vector indicating whic type of normalization has to be used. \code{Type = "1"} scales
the vector into the range [-1, 1] and \code{Type = "2"} scales the vector into the range [0, 1].}
}
\value{
the normalized vector
}
\description{
Min-max scaling to normalize a vector
}
\examples{
x <- rnorm(10)
normalize(x)
}
