% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{print.BiRankFr}
\alias{print.BiRankFr}
\title{Print method for object of type BiRankFr}
\usage{
\method{print}{BiRankFr}(x, ...)
}
\arguments{
\item{x}{Object of type BiRankFr}

\item{...}{Ellipses argument, to be passed to \code{print} function.}
}
\value{
Nothing
}
\description{
Print method for object of type BiRankFr
}
