% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnGenerator.R
\name{print.sfnData}
\alias{print.sfnData}
\title{Print function for a sfnData object}
\usage{
\method{print}{sfnData}(x, ...)
}
\arguments{
\item{x}{an object of class \code{sfnData}, resulting from \code{\link{sfnGenerator}}}

\item{...}{arguments passed to \code{\link{print}}}
}
\value{
The original \code{sfnData} object is returned.
}
\description{
Prints the call, target class imbalance, class imbalance in the synthetic data set, the first part of the data set and
the summary ggplot object.
}
\seealso{
[sfnGenerator()]
}
